cleanup mod imports and types
