+++ src/client/bandwidth_test.rs
@@ -1,0 +1,7 @@
+
+
+pub fn bandwidth_test() {
+
+    let min_bandwidth = 1000; // 1KB/s
+
+}
