+++ src/client/mod.rs
@@ -1,9 +1,9 @@
 mod start_client_flow;
 mod select_relay_algo;
-mod build_circuit;
+mod circuit;
 mod payments_ledger;
 
-pub use start_client_flow::{start_client_flow};
-pub use select_relay_algo::{simple_relay_selection_algo};
-pub use build_circuit::{build_circuit, pregen_extend_paid_circuit_hashes};
-pub use payments_ledger::{init_payments_ledger};
+pub use start_client_flow::*;
+// pub use select_relay_algo::*;
+// pub use circuit::*;
+// pub use payments_ledger::*;
