+++ src/client/payments_loop.rs
@@ -1,0 +1,18 @@
+pub fn start_payments_loop () {
+
+    // is round expired
+    // if no then do bandwidth test
+    // if good then pay relay
+    // wait for next round
+
+}
+
+
+
+pub fn is_round_expired () {
+
+}
+
+pub fn kill_circuit () {
+
+}
