+++ src/rpc/mod.rs
@@ -1,13 +1,11 @@
 mod get_relay_descriptors;
 mod rpc_client;
 mod get_current_consensus;
-mod utils;
 mod get_conf;
 mod extend_paid_circuit;
 
-pub use rpc_client::{rpc_client, RpcConfig};
-pub use get_relay_descriptors::{get_relay_descriptors, Relay};
-pub use get_current_consensus::{get_current_consensus, ConsensusRelay, RelayTag};
-pub use utils::microdesc_to_fingerprint;
-pub use get_conf::{get_conf, get_conf_payment_circuit_max_fee};
-pub use extend_paid_circuit::extend_paid_circuit;
+pub use rpc_client::*;
+pub use get_relay_descriptors::*;
+pub use get_current_consensus::*;
+pub use get_conf::*;
+pub use extend_paid_circuit::*;
